chore: extract Composer repo operations into composer-repo.ts
